Capture and store email address

+++ lib/GrotSpot/Controller/Ajax.pm
@@ -4,6 +4,18 @@ use strict;
 use warnings;
 use parent 'Catalyst::Controller';
 
+use Email::Valid;
+
+sub auto : Private {
+    my ( $self, $c ) = @_;
+
+    use Data::Dumper;
+    local $Data::Dumper::Sortkeys = 1;
+    warn Dumper( $c->session );
+
+    1;
+}
+
 sub store_rating : Local {
     my ( $self, $c ) = @_;
 
@@ -23,11 +35,15 @@ sub store_rating : Local {
       $c->future_discount( $c->session->{ratings_stored} );
     my $session_id = $c->sessionid;
 
+    # get the email if we have one
+    my $email_id = $c->session->{email_id} || undef;
+
     # save the rating
     my $rating = $location->add_to_ratings(
         {
             score      => $score,
-            session_id => $session_id
+            session_id => $session_id,
+            email_id   => $email_id,
         }
     ) || die "Could not save rating";
 
@@ -46,9 +62,40 @@ sub store_rating : Local {
     };
 }
 
+sub store_email : Local {
+    my ( $self, $c ) = @_;
+
+    my $req       = $c->req;
+    my $raw_email = $req->param('email');
+
+    # check that the email is valid
+    my $valid_email = Email::Valid->address( lc $raw_email );
+    if ( !$valid_email ) {
+        $c->response->body("Bad email address");
+        $c->response->status(404);
+        $c->detach;
+    }
+
+    # find the email in db
+    my $email =
+      $c->model('DB::Email')->find_or_create( { email => $valid_email } )
+      || die "could not create entry for $valid_email";
+
+    # set the email into the session
+    $c->session->{email}    = $valid_email;
+    $c->session->{email_id} = $email->id;
+
+    # find all ratings done in this session with no email and fix them.
+    $c->model('DB::Rating')
+      ->search( { session_id => $c->sessionid, email_id => undef } )
+      ->update( { email_id => $email->id } );
+
+    $c->stash->{json_data} = { email_saved => 1, };
+}
+
 sub end : Private {
     my ( $self, $c ) = @_;
-    $c->detach('View::JSON');
+    $c->detach('View::JSON') unless $c->res->body;
 }
 
 1;
